Move annotate submodules up a level.
