+++ nimare/annotate/gclda.py
@@ -9,10 +9,10 @@ import pandas as pd
 import nibabel as nib
 from scipy.stats import multivariate_normal
 
-from ...due import due
-from ...base import NiMAREBase
-from ...utils import get_template
-from ... import references
+from ..due import due
+from ..base import NiMAREBase
+from ..utils import get_template
+from .. import references
 
 LGR = logging.getLogger(__name__)
