--- nimare/annotate/topic/__init__.py
@@ -1,9 +1,0 @@
-"""
-Automated annotation with text-derived topic models.
-"""
-from .boltzmann import BoltzmannModel
-from .gclda import GCLDAModel
-# from .glove import GloveModel
-from .lda import LDAModel
-
-__all__ = ['BoltzmannModel', 'GCLDAModel', 'LDAModel']  # , 'GloveModel']
