--- nimare/annotate/vector/__init__.py
@@ -1,7 +1,0 @@
-"""
-Automated annotation with text-derived vector models.
-"""
-from .word2brain import Word2BrainModel
-from .text2brain import Text2BrainModel
-
-__all__ = ['Word2BrainModel', 'Text2BrainModel']
